requirements.txt
azure-storage-blob==1.4.0
azure-storage-common==1.4.0
PyPDF2==1.26.0